dtops finalizados

+++ TP1/Ejercicio_1/src/main/java/com/ejemploSpring/Ejercicio/dtos/detallePedido/DetallePedidoDto.java
@@ -2,10 +2,12 @@ package com.ejemploSpring.Ejercicio.dtos.detallePedido;
 
 import com.ejemploSpring.Ejercicio.entities.DetallePedido;
 
+import java.math.BigDecimal;
+
 public record DetallePedidoDto(
         Long id,
         int cantidad,
-        Double subtotal,
+        BigDecimal subtotal,
         Long productoId,
         String productoNombre,
         Long pedidoId
